requirements.txt
fastapi
uvicorn
motor
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator